requirements.txt
streamlit
pandas
scikit-learn
matplotlib
pillow
joblib